WIP - CookieControl as external component
